Move deploy command from rivet version deploy to rivet deploy
