+++ cli/src/commands/version.rs
@@ -3,15 +3,11 @@ use std::collections::HashSet;
 use anyhow::{bail, Context, Error, Result};
 use clap::Parser;
 use cli_core::rivet_api::{apis, models};
-use serde::{Deserialize, Serialize};
-use serde_json::json;
+use serde::Deserialize;
 use tabled::Tabled;
 use uuid::Uuid;
 
-use crate::{
-	commands::{cdn, docker},
-	util::{gen, struct_fmt, term},
-};
+use crate::{commands::deploy, util::term};
 
 #[derive(Parser)]
 pub enum SubCommand {
@@ -23,10 +19,6 @@ pub enum SubCommand {
 		/// The version ID
 		version: String,
 	},
-
-	/// Pushes the build and site and creates a new version
-	#[clap(alias = "push-and-create", alias = "create", alias = "publish")]
-	Deploy(DeployOpts),
 
 	/// Returns the config for a version
 	#[clap(alias = "read-config")]
@@ -43,6 +35,17 @@ pub enum SubCommand {
 	/// Show the a version's dashboard
 	#[clap(alias = "dashboard", alias("dash"))]
 	View { version: String },
+
+	/// Deprecated. Use `rivet deploy`.
+	///
+	/// Pushes the build and site and creates a new version
+	#[clap(
+		hide = true,
+		alias = "push-and-create",
+		alias = "create",
+		alias = "publish"
+	)]
+	Deploy(deploy::Opts),
 }
 
 impl SubCommand {
@@ -154,80 +157,6 @@ impl SubCommand {
 				Ok(())
 			}
 		}
-	}
-}
-
-#[derive(Parser)]
-pub struct DeployOpts {
-	/// Name of the version to create
-	#[clap(long = "name", alias = "display-name")]
-	display_name: Option<String>,
-
-	/// Override specific properties of the config
-	#[clap(long = "override", short)]
-	overrides: Vec<String>,
-
-	/// Namespace ID to deploy to
-	#[clap(short = 'n', long)]
-	namespace: Option<String>,
-
-	/// Deprecated.
-	///
-	/// The build tag to upload
-	#[clap(hide = true, long)]
-	build_tag: Option<String>,
-
-	/// Deprecated.
-	///
-	/// The name to assign to the build
-	#[clap(hide = true, long)]
-	build_name: Option<String>,
-
-	/// Deprecated.
-	///
-	/// The path to the site directory to upload
-	#[clap(hide = true, long)]
-	site_path: Option<String>,
-
-	/// Deprecated.
-	///
-	/// The name of the site that will be created
-	#[clap(hide = true, long)]
-	site_name: Option<String>,
-
-	#[clap(long, value_parser)]
-	format: Option<struct_fmt::Format>,
-}
-
-impl DeployOpts {
-	pub async fn execute(&self, ctx: &cli_core::Ctx) -> Result<()> {
-		// Parse overrides
-		let mut overrides = parse_config_override_args(&self.overrides)?;
-
-		// Build & push site & build before creating version
-		build_and_push_compat(
-			ctx,
-			&mut overrides,
-			&self.build_tag,
-			&self.build_name,
-			&self.site_path,
-			&self.site_name,
-			&self.format,
-		)
-		.await?;
-
-		// Create version
-		let output = create(
-			ctx,
-			self.display_name.as_ref().map(String::as_str),
-			overrides,
-			self.namespace.as_ref().map(String::as_str),
-			self.format.as_ref(),
-		)
-		.await?;
-		struct_fmt::print_opt(self.format.as_ref(), &output)?;
-
-		Ok(())
 	}
 }
 
@@ -398,129 +327,6 @@ pub fn build_mock_config_dependencies(version: &mut models::CloudVersionConfig) 
 	Ok(())
 }
 
-/// Builds the Docker image and CDN site if needed.
-pub async fn build_config_dependencies(
-	ctx: &cli_core::Ctx,
-	version: &mut models::CloudVersionConfig,
-	display_name: &str,
-	format: Option<&struct_fmt::Format>,
-) -> Result<()> {
-	if let Some(matchmaker) = version.matchmaker.as_mut() {
-		// matchmaker.docker
-		let default_image_id = if let Some(docker) = matchmaker.docker.as_mut() {
-			let image_id = build_and_push_image(ctx, display_name, docker, format, None).await?;
-			docker.image_id = image_id;
-			image_id
-		} else {
-			None
-		};
-
-		// matchmaker.game_modes.*.docker
-		if let Some(game_modes) = matchmaker.game_modes.as_mut() {
-			for (_, game_mode) in game_modes.iter_mut() {
-				if let Some(docker) = game_mode.docker.as_mut() {
-					docker.image_id =
-						build_and_push_image(ctx, display_name, docker, format, default_image_id)
-							.await?;
-				}
-			}
-		}
-	}
-
-	// Build CDN
-	if let Some(cdn) = version.cdn.as_mut() {
-		build_and_push_site(ctx, display_name, cdn, format).await?;
-	}
-
-	Ok(())
-}
-
-/// Builds image if not specified and returns the image ID.
-///
-/// The image ID is chosen in order of priority:
-///
-/// - `dockerfile` Build the Dockerfile
-/// - `image` Upload a prebuilt image
-/// - `default_image_id` Use the image ID defined at the base of the matchmaker config
-///
-/// If none are true, `None` is returned.
-pub async fn build_and_push_image(
-	ctx: &cli_core::Ctx,
-	display_name: &str,
-	docker: &mut Box<models::CloudVersionMatchmakerGameModeRuntimeDocker>,
-	format: Option<&struct_fmt::Format>,
-	default_image_id: Option<Uuid>,
-) -> Result<Option<Uuid>> {
-	if docker.image_id.is_none() {
-		if let Some(dockerfile) = &docker.dockerfile {
-			let push_output = docker::build_and_push(
-				ctx,
-				&docker::BuildPushOpts {
-					dockerfile: dockerfile.clone(),
-					name: Some(display_name.to_string()),
-					format: format.cloned(),
-				},
-			)
-			.await?;
-
-			return Ok(Some(push_output.image_id));
-		} else if let Some(docker_image) = docker.image.as_ref() {
-			let push_output = docker::push(
-				ctx,
-				&docker::PushOpts {
-					tag: docker_image.clone(),
-					name: Some(display_name.to_string()),
-					format: format.cloned(),
-				},
-			)
-			.await?;
-
-			return Ok(Some(push_output.image_id));
-		} else if let Some(image_id) = default_image_id {
-			return Ok(Some(image_id));
-		}
-	}
-
-	Ok(None)
-}
-pub async fn build_and_push_site(
-	ctx: &cli_core::Ctx,
-	display_name: &str,
-	cdn: &mut Box<models::CloudVersionCdnConfig>,
-	format: Option<&struct_fmt::Format>,
-) -> Result<()> {
-	if cdn.site_id.is_none() {
-		if let Some(build_output) = &cdn.build_output {
-			if let Some(build_command) = &cdn.build_command {
-				let push_output = cdn::build_and_push(
-					ctx,
-					&cdn::BuildPushOpts {
-						command: build_command.clone(),
-						path: build_output.clone(),
-						name: Some(display_name.to_string()),
-						format: format.cloned(),
-					},
-				)
-				.await?;
-				cdn.site_id = Some(push_output.site_id);
-			} else {
-				let push_output = cdn::push(
-					ctx,
-					&cdn::PushOpts {
-						path: build_output.clone(),
-						name: Some(display_name.to_string()),
-						format: format.cloned(),
-					},
-				)
-				.await?;
-				cdn.site_id = Some(push_output.site_id);
-			}
-		}
-	}
-
-	Ok(())
-}
-
 pub fn dashboard_url(game_id: &str, version_id: &str) -> String {
 	format!(
 		"https://hub.rivet.gg/developer/games/{game_id}/versions/{version_id}",
@@ -535,165 +341,4 @@ pub fn rivet_game_url(game_name_id: &str, namespace_name_id: &str) -> String {
 	} else {
 		format!("https://{game_name_id}--{namespace_name_id}.rivet.game/")
 	}
-}
-
-#[derive(Serialize)]
-pub struct CreateOutput {
-	pub version_id: Uuid,
-}
-
-/// Creates a new Rivet version.
-pub async fn create(
-	ctx: &cli_core::Ctx,
-	display_name: Option<&str>,
-	overrides: Vec<(String, serde_json::Value)>,
-	namespace_name_id: Option<&str>,
-	format: Option<&struct_fmt::Format>,
-) -> Result<CreateOutput> {
-	// Fetch game data
-	let game_res = apis::cloud_games_games_api::cloud_games_games_get_game_by_id(
-		&ctx.openapi_config_cloud,
-		&ctx.game_id,
-		None,
-	)
-	.await;
-	if let Err(err) = game_res.as_ref() {
-		println!("Error: {err:?}");
-	}
-	let game_res = game_res.context("cloud_games_games_get_game_by_id")?;
-	let namespace = if let Some(namespace) = namespace_name_id {
-		Some(
-			game_res
-				.game
-				.namespaces
-				.iter()
-				.find(|x| x.name_id == namespace)
-				.context("namespace not found")?,
-		)
-	} else {
-		None
-	};
-
-	// Generate version name if needed
-	let display_name = if let Some(x) = &display_name {
-		x.to_string()
-	} else {
-		gen::version_display_name(&game_res.game)?
-	};
-
-	// Parse config
-	let mut rivet_config = read_config(overrides, namespace_name_id).await?;
-	build_config_dependencies(ctx, &mut rivet_config, &display_name, format).await?;
-
-	// Create game version
-	let version_res = apis::cloud_games_versions_api::cloud_games_versions_create_game_version(
-		&ctx.openapi_config_cloud,
-		&ctx.game_id,
-		models::CloudGamesCreateGameVersionRequest {
-			display_name: display_name.clone(),
-			config: Box::new(rivet_config),
-		},
-	)
-	.await;
-	if let Err(err) = version_res.as_ref() {
-		println!("Error: {err:?}");
-	}
-	let version_res = version_res.context("versions_create_game_version")?;
-	let version_id = version_res.version_id;
-
-	eprintln!();
-	term::status::success("Deployed Version", &display_name);
-	term::status::info(
-		"Version Dashboard",
-		dashboard_url(&ctx.game_id, &version_id.to_string()),
-	);
-
-	// Deploy to namespace
-	if let Some(namespace) = namespace {
-		eprintln!();
-		term::status::info(
-			"Deploying to Namespace",
-			format!("{} -> {}", display_name, namespace.display_name),
-		);
-		let update_version_res =
-			apis::cloud_games_namespaces_api::cloud_games_namespaces_update_game_namespace_version(
-				&ctx.openapi_config_cloud,
-				&ctx.game_id,
-				&namespace.namespace_id.to_string(),
-				models::CloudGamesNamespacesUpdateGameNamespaceVersionRequest {
-					version_id: version_id.clone(),
-				},
-			)
-			.await;
-		if let Err(err) = update_version_res.as_ref() {
-			println!("Error: {err:?}");
-		}
-		update_version_res.context("cloud_games_namespaces_update_game_namespace_version")?;
-		term::status::success(
-			"Deploy Succeeded",
-			rivet_game_url(&game_res.game.name_id, &namespace.name_id),
-		);
-	}
-
-	eprintln!();
-
-	Ok(CreateOutput { version_id })
-}
-
-/// Backwards compatibility for site & Docker build pushing from the CLI flags.
-///
-/// Developers should override config properties instead. For example: `rivet deploy -o matchmaker.docker.image_id=xxxx -o
-/// cdn.path=xxxx`
-pub async fn build_and_push_compat(
-	ctx: &cli_core::Ctx,
-	overrides: &mut Vec<(String, serde_json::Value)>,
-	build_tag: &Option<String>,
-	build_name: &Option<String>,
-	site_path: &Option<String>,
-	site_name: &Option<String>,
-	format: &Option<struct_fmt::Format>,
-) -> Result<()> {
-	let site_output = if let Some(site_path) = site_path {
-		Some(
-			cdn::push(
-				ctx,
-				&cdn::PushOpts {
-					path: site_path.clone(),
-					name: site_name.clone(),
-					format: format.clone(),
-				},
-			)
-			.await?,
-		)
-	} else {
-		None
-	};
-
-	let build_output = if let Some(build_tag) = build_tag {
-		Some(
-			docker::push(
-				ctx,
-				&docker::PushOpts {
-					tag: build_tag.clone(),
-					name: build_name.clone(),
-					format: format.clone(),
-				},
-			)
-			.await?,
-		)
-	} else {
-		None
-	};
-
-	if let Some(site_output) = site_output {
-		overrides.push(("cdn.site_id".into(), json!(site_output.site_id)));
-	}
-	if let Some(build_output) = build_output {
-		overrides.push((
-			"matchmaker.docker.image_id".into(),
-			json!(build_output.image_id),
-		));
-	}
-
-	Ok(())
 }
